Moves distinct and partition to util.py
